fix(imu): String(float,dp) неоднозначность — dp как unsigned int (ESP32 core 3.3)
uint8_t dp точно матчил unsigned char base у целочисленных перегрузок String
и одновременно float-перегрузку → ambiguous. unsigned int снимает неоднозначность.

+++ xiao_cam_stream/xiao_imu.h
@@ -180,7 +180,7 @@ static inline bool xiaoImuIsOk() { return gImuOk; }
 static inline float xiaoImuYaw() { return gImuYaw; }
 
 static inline void xiaoImuAppendTelemetry(String &j, bool &comma) {
-  auto add = [&](const char *k, float v, uint8_t dp) {
+  auto add = [&](const char *k, float v, unsigned int dp) {
     if (comma) j += ',';
     comma = true;
     j += '"'; j += k; j += "\":"; j += String(v, dp);
